feat: expose wizards as web components
