+++ package.json
@@ -34,6 +34,7 @@
 		"prettier": "^3.6.2",
 		"react": "^19.1.1",
 		"react-dom": "^19.1.1",
+		"react-to-webcomponent": "^2.0.1",
 		"tailwindcss": "^4.1.12",
 		"typescript": "^5.9.2",
 		"typescript-eslint": "^8.41.0",
