+++ public/test-install-esp-bridge.html
@@ -1,0 +1,15 @@
+<!doctype html>
+<html lang="en">
+	<head>
+		<meta charset="UTF-8" />
+		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
+		<title>Install ESP Bridge Firmware</title>
+	</head>
+	<body>
+		<!-- The web component will be rendered here -->
+		<install-esp-bridge-firmware></install-esp-bridge-firmware>
+
+		<!-- Load the web component script -->
+		<script type="module" src="./standalone/install-esp-bridge-firmware.js"></script>
+	</body>
+</html>
