+++ public/test-install-esphome.html
@@ -1,0 +1,18 @@
+<!doctype html>
+<html lang="en">
+	<head>
+		<meta charset="UTF-8" />
+		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
+		<title>Install ESPHome Firmware</title>
+	</head>
+	<body>
+		<!-- The web component will be rendered here -->
+		<install-esphome-firmware></install-esphome-firmware>
+
+		<!-- Load the web component script -->
+		<script
+			type="module"
+			src="./standalone/install-esphome-firmware.js"
+		></script>
+	</body>
+</html>
