+++ src/wizards/update-esp-firmware/index.ts
@@ -1,2 +1,2 @@
-export { updateESPFirmwareWizardConfig, updateESPBridgeWizardConfig, updateESPHomeWizardConfig } from './wizard';
+export { updateESPFirmwareWizardConfig, installESPBridgeFirmwareWizardConfig as updateESPBridgeWizardConfig, installESPHomeFirmwareWizardConfig as updateESPHomeWizardConfig } from './wizard';
 export type { UpdateESPFirmwareState, ESPFirmwareOption, ConfigureState } from './wizard';
